+++ src/Guid/Guid.php
@@ -37,20 +37,13 @@ use Ramsey\Uuid\UuidInterface;
  */
 class Guid extends Uuid implements UuidInterface
 {
-    /**
-     * @param string[] $fields
-     */
     public function __construct(
-        array $fields,
+        Fields $fields,
         NumberConverterInterface $numberConverter,
         CodecInterface $codec,
         TimeConverterInterface $timeConverter,
         CalculatorInterface $calculator
     ) {
-        $this->fields = new Fields((string) hex2bin(implode('', $fields)));
-        $this->codec = $codec;
-        $this->numberConverter = $numberConverter;
-        $this->timeConverter = $timeConverter;
-        $this->calculator = $calculator;
+        parent::__construct($fields, $numberConverter, $codec, $timeConverter, $calculator);
     }
 }